Move the trimesh-only helpers on to skrobot.utils.mesh
Vertex normals, transparency and material colors are mesh operations with
no URDF and no filesystem in them, which is what skrobot.utils.mesh already
holds.  They are imported inside the functions that use them, so neither
urdf.py nor robot_model.py starts pulling in trimesh at import time.
